// types.ts
export enum NoteType {
  SEMIBREVE = 'SEMIBREVE',
  DOTTED_MINIM = 'DOTTED_MINIM',
  MINIM = 'MINIM',
  DOTTED_CROTCHET_QUAVER = 'DOTTED_CROTCHET_QUAVER',
  CROTCHET = 'CROTCHET',
  TWO_QUAVERS = 'TWO_QUAVERS',
  FOUR_SEMIQUAVERS = 'FOUR_SEMIQUAVERS',
  QUAVER_TWO_SEMIQUAVERS = 'QUAVER_TWO_SEMIQUAVERS',
  TWO_SEMIQUAVERS_QUAVER = 'TWO_SEMIQUAVERS_QUAVER',
  DOTTED_QUAVER_SEMIQUAVER = 'DOTTED_QUAVER_SEMIQUAVER',
  SEMIQUAVER_DOTTED_QUAVER = 'SEMIQUAVER_DOTTED_QUAVER',
  TRIPLET_QUAVERS = 'TRIPLET_QUAVERS',
  UNKNOWN = 'UNKNOWN',

  // Rests
  SEMIBREVE_REST = 'SEMIBREVE_REST',
  DOTTED_MINIM_REST = 'DOTTED_MINIM_REST',
  MINIM_REST = 'MINIM_REST',
  CROTCHET_REST = 'CROTCHET_REST',
  DOTTED_CROTCHET_QUAVER_REST = 'DOTTED_CROTCHET_QUAVER_REST',
  TWO_QUAVERS_REST = 'TWO_QUAVERS_REST',
  FOUR_SEMIQUAVERS_REST = 'FOUR_SEMIQUAVERS_REST',
  QUAVER_TWO_SEMIQUAVERS_REST = 'QUAVER_TWO_SEMIQUAVERS_REST',
  TWO_SEMIQUAVERS_QUAVER_REST = 'TWO_SEMIQUAVERS_QUAVER_REST',
  DOTTED_QUAVER_SEMIQUAVER_REST = 'DOTTED_QUAVER_SEMIQUAVER_REST',
  SEMIQUAVER_DOTTED_QUAVER_REST = 'SEMIQUAVER_DOTTED_QUAVER_REST',
  TRIPLET_QUAVERS_REST = 'TRIPLET_QUAVERS_REST',
}

export interface RhythmicNote {
  lyric: string;
  note: NoteType;
}